# [WIP] Refactor solve()