… sympy/solvers/tests/test_solvers.py