@sayandip18 2021-01-15
```
sympy/solvers/tests/test_solvers.py::test_solve_args - Took 22.401 seconds
________________________________________________________________________________
_____________ sympy/solvers/tests/test_solvers.py:test_solve_args ______________
Traceback (most recent call last):
  File "/home/sayandip/sympy/sympy/solvers/tests/test_solvers.py", line 167, in test_solve_args
    assert solve(Eq(x**2, 0.0)) == [0]  # issue 19048
  File "/home/sayandip/sympy/sympy/solvers/solvers.py", line 907, in solve
    '''))
NotImplementedError: 
Unanticipated argument of Eq when other arg is True or False.

```
This failing test will pass after #20801 is merged