@oscarbenjamin 2021-01-18
Thanks for taking this on. I think it's important that we move as much as possible of the logic outside of the main `solve` function which is otherwise far too big and complicated.

It probably makes sense to move all the nested functions near to the top of the function. They should have docstrings/comments that explain what they do.

There should be other comments with examples to make it clear what each early exit does.

More generally I think that a number of these code paths should be deprecated and removed but that shouldn't be part of this PR. Comments should be added though so it is clear what each part does. That will help when thinking about removing or deprecating things later.