@sayandip18 2021-01-18
Seconded. Also, I think it's important that `_solve_system()` has a docstring too, when it's the cse that `_solve()` has one. But I guess that should be in a different PR.