@oscarbenjamin 2021-02-06
I made various changes in #18844 since I stopped working on #17793. I think that the way you have brought these changes forwards from #17793 has undone some of those.

In particular I made some subtle changes to different parts of `solve` to make it faster for large sparse systems of equations. Undoing those is leading to the benchmark changes that you have noticed.

Here is an example:
```python
from sympy import *

def make_system(n):
    xs = symbols(f'x:{n}')
    ys = symbols(f'y:{n}')
    eqs = []
    for x, y in zip(xs, ys):
        eqs.append(Eq(x + y, 1))
        eqs.append(Eq(x - y, 0))
    return eqs, xs + ys
```
On master I get:
```python
In [3]: eqs, syms = make_system(100)

In [4]: %time ok = solve(eqs, syms)
CPU times: user 816 ms, sys: 6.68 ms, total: 822 ms
Wall time: 823 ms
```
Whereas with the PR I get:
```python
In [2]: eqs, syms = make_system(100)

In [3]: %time ok = solve(eqs, syms)
CPU times: user 3.37 s, sys: 16.5 ms, total: 3.39 s
Wall time: 3.4 s
```
If I profile that then we can see what the major slowdown is:
```python
In [5]: %prun -s cumulative ok = solve(eqs, syms)

         11983909 function calls (10915113 primitive calls) in 11.122 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000   11.122   11.122 {built-in method builtins.exec}
        1    0.000    0.000   11.122   11.122 <string>:1(<module>)
        1    0.012    0.012   11.122   11.122 solvers.py:398(solve)
137608/135708    0.288    0.000    9.003    0.000 cache.py:69(wrapper)
141845/6905    0.351    0.000    8.915    0.001 {built-in method builtins.any}
    41201    0.052    0.000    8.432    0.000 basic.py:1167(has)
   135741    0.142    0.000    8.310    0.000 basic.py:1214(<genexpr>)
    94540    0.728    0.000    8.168    0.000 basic.py:1216(_has)
    40200    0.132    0.000    7.218    0.000 solvers.py:1035(<genexpr>)
   563436    0.664    0.000    6.642    0.000 basic.py:1232(<genexpr>)
   469296    0.435    0.000    3.406    0.000 basic.py:1238(<lambda>)
477796/477496    1.224    0.000    3.047    0.000 expr.py:137(__eq__)
   570541    0.338    0.000    2.606    0.000 basic.py:2040(__next__)
   571741    0.307    0.000    2.268    0.000 {built-in method builtins.next}
```
What is slow here is the call to `has(*symset)`. Why is that slow? Well we have N equations and N symbols. For each equation we call `eq.has(*symset)`. Internally `has` is very slow and also loops over its arguments calling `_has` for each:
https://github.com/sympy/sympy/blob/dc962d75fa576051aa9306bfd90419f5c80ed657/sympy/core/basic.py#L1214
That means that `_has` gets called `N^2` times making the whole operation quadratic.

Changing it in solve from `eq.has(*symset)` to `eq.free_symbols & symset` is more efficient as it only walks once through the expression tree of `eq` to collect the symbols and then computes a single set intersection. Intersecting a small set with a large set is efficient in Python and has time proportional to the small set meaning that the time for this operation depends on the number of symbols in `eq` but not on the number of symbols in `symset`. That makes it efficient for a large sparse system as above.