@oscarbenjamin 2021-02-17
Have you profiled this?

Besides the benchmarks I showed a relevant case for comparison above: https://github.com/sympy/sympy/pull/20790#issuecomment-774469441