@sayandip18 2021-02-18
I ran the same benchmark you used in the comment.
On master, I'm getting this:
```
In [5]: %time ok = solve(eqs, syms)
CPU times: user 988 ms, sys: 644 µs, total: 989 ms
Wall time: 1.01 s

```
On this PR, I'm getting this:
```
In [4]: %time ok = solve(eqs, syms)
CPU times: user 966 ms, sys: 5.71 ms, total: 971 ms
Wall time: 1 s

```

I ran [this](https://github.com/sympy/sympy/pull/18844#issuecomment-598260831) benchmark too and this is what I found:

On master
```
In [2]: %time ok=dosolve(10)
CPU times: user 383 ms, sys: 7.31 ms, total: 391 ms
Wall time: 1.01 s

In [3]: %time ok=dosolve(20)
CPU times: user 1.13 s, sys: 22.4 ms, total: 1.15 s
Wall time: 1.16 s


``` 
On this PR
```
In [10]: %time ok=dosolve(10)
CPU times: user 311 ms, sys: 0 ns, total: 311 ms
Wall time: 314 ms

In [3]: %time ok=dosolve(20)
CPU times: user 1.13 s, sys: 7.43 ms, total: 1.14 s
Wall time: 1.15 s
```