@oscarbenjamin 2021-02-18
Try this:
```python
from sympy import *

def make_system(n):
    xs = symbols(f'x:{n}')
    ys = symbols(f'y:{n}')
    eqs = []
    for x, y in zip(xs, ys):
        eqs.append(Eq(x + y, 1))
        eqs.append(Eq(x - y, 0))
    z = Symbol('z')
    eqs.append(re(z))
    return eqs, xs + ys + (z,)
```
Then
```
In [2]: eqs, syms = make_system(100)

In [3]: %time ok = solve(eqs, syms)
CPU times: user 2.98 s, sys: 8.27 ms, total: 2.99 s
Wall time: 2.99 s
```
That's 3x slower than master.