@sayandip18 2021-02-18
On master I'm getting this 
```
In [3]: %time ok = solve(eqs, syms)
CPU times: user 1.28 s, sys: 431 µs, total: 1.28 s
Wall time: 1.29 s
```
whereas on the PR
```
In [7]: %time ok = solve(eqs, syms)
CPU times: user 1.26 s, sys: 0 ns, total: 1.26 s
Wall time: 1.26 s
```