@oscarbenjamin 2021-02-26
This is very incomplete compared to what I was aiming for in #17993. I really wanted to make it so that the steps in the solve function were easier to see and better documented. This moves a few things around but it's kind of a half-refactoring so I'm not sure what it actually achieves.