@sayandip18 2021-02-27
Thanks for the feedback, I'll try to modify solve() a bit more